Integrated Condition into Flags
The reasoning behind this is that conditional are inherently tied to the event flag and their creation as scriptable objects felt like an unnecessary middle-man to the process - conditionals were rarely re-used outside of maybe same events triggers, which were not something that just having the unlock flags handle themselves could not handle. If I made an event, I wouldn't want to have to made a conditional just to use it in an unlock flag. The way it is now, where a conditional is serialized as part of an unlock flag, makes more sense since it's more direct. After all, for most use cases, the number of actual objects involved was just +1 for condition.